+++ cpp/src/arrow/CMakeLists.txt
@@ -180,7 +180,6 @@ set(ARROW_SRCS
     util/key_value_metadata.cc
     util/memory.cc
     util/parsing.cc
-    util/receiver.cc
     util/string.cc
     util/string_builder.cc
     util/task_group.cc
